+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog follow https://keepachangelog.com/ format.
 *   `epy`:
     *   Add `epy.is_test` to check whether we're running in a test environment.
     *   Add `epy.typing.Json`.
+    *   Add `epy.ExitStack` which allows setting the contextmanagers during init.
 *   `exm`:
     *   Add `exm.url_to_python_only_logs()` and `exm.curr_job_name()` to add
         artifact to Python only logs (without all the verbose C++ junk ).